backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
elasticsearch==8.10.0
kafka-python==2.0.2
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0